SERVER_BACKEND/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
gunicorn==23.0.0
idna==3.7
nh3==0.2.17
packaging==24.1
# psycopg2==2.9.9
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.2
